requirements.txt
django-flexible-content
